requirements.txt
fastapi
uvicorn
gunicorn
httpx
openai
newspaper3k
lxml_html_clean
pypdf
playwright
